workflows/cloudrun_to_cloudrun/functions/get_from_gcs/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
ptvsd==4.3.2 # Required for debugging.
requests==2.25.1
google-cloud-storage==2.1.0
